use android test because roboelectric tests are failing
